:recycle: Move the assertions into their own directory
