+++ src/client/requests.rs
@@ -58,3 +58,4 @@ impl From<Url> for MockReq {
         }
     }
 }
+
